added separate endpoint for inviting users, amended validation to validate Profile model and Invitation model - use dummy uuid for invitation validate as profile hasn't yet been created
