+++ lib/routers/establishment.js
@@ -51,5 +51,6 @@ router.use('/:establishment/roles', require('./roles'));
 router.use('/:establishment/place(s)?', require('./places'));
 router.use('/:establishment/profile(s)?', require('./profile'));
 router.use('/:establishment/project(s)?', require('./projects'));
+router.use('/:establishment/invite-user', require('./invite-user'));
 
 module.exports = router;
