+++ lib/routers/profile.js
@@ -1,41 +1,8 @@
 const { Router } = require('express');
 const isUUID = require('uuid-validate');
 const { NotFoundError } = require('../errors');
-const permissions = require('../middleware/permissions');
 
 const router = Router({ mergeParams: true });
-
-const submit = (action) => {
-  return (req, res, next) => {
-    const params = {
-      action,
-      model: 'invitation',
-      data: { ...req.body, establishment: req.establishment.id },
-      id: res.profile && res.profile.id
-    };
-    req.workflow(params)
-      .then(response => {
-        res.response = response;
-        next();
-      })
-      .catch(next);
-  };
-};
-
-const validateSchema = () => {
-  return (req, res, next) => {
-    let data = { ...req.body, establishmentId: req.establishment.id };
-
-    if (res.profile) {
-      data = Object.assign({}, res.place, data);
-    }
-
-    const { Profile } = req.models;
-    const error = Profile.validate(data);
-
-    return error ? next(error) : next();
-  };
-};
 
 router.param('id', (req, res, next, id) => {
   if (!isUUID(id)) {
@@ -91,7 +58,5 @@ router.get('/:id', (req, res, next) => {
   res.response = res.profile;
   next();
 });
-
-router.post('/', permissions('profile.invite'), validateSchema(), submit('create'));
 
 module.exports = router;
